clean up code

+++ jsrc/jdlllic.c
@@ -9,6 +9,7 @@
 
 #include <stdlib.h>
 #include <math.h>
+#include <ctype.h>
 
 #if !SY_WINCE
 #include <time.h>
@@ -200,7 +201,7 @@ F2(jtunlock2){int i,j,len,tlen;UC c1,c2,k1[SK],*lp,*sp,*d;
  tlen=(int)AN(w);
  if(!tlen || 255!=d[0] || 0 != d[1] || tlen<8+SK) return w;	/* not jl */
  memcpy(k1, d+8, SK);
- for(i=0;i<sizeof(int);++i)	*(i+(UC*)&len) = k1[i] ^ d[4+i];
+ for(i=0;i<(int)sizeof(int);++i)	*(i+(UC*)&len) = k1[i] ^ d[4+i];
 #if !(SYS & SYS_LILENDIAN)
  len=swapint(len);
 #endif
